requirements.txt
absl-py==0.7.1
astor==0.8.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
Keras==2.1.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
nose==1.3.7
numpy==1.17.0
Pillow==6.1.0
protobuf==3.9.0
PyYAML==5.1.2
scipy==1.3.0
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==0.15.5
wrapt==1.11.2
